# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DataNova API - Plateforme Veille Réglementaire
  description: API REST pour la gestion de la veille réglementaire Hutchinson SA
  version: 1.0.0
  contact:
    name: Backend Team
    email: khadidja@hutchinson.com

servers:
  - url: http://localhost:8000/api
    description: Serveur de développement

paths:
  /regulations:
    get:
      summary: Liste des réglementations
      description: Récupère toutes les réglementations avec filtres optionnels
      tags:
        - Réglementations
      parameters:
        - name: status
          in: query
          description: Filtrer par statut
          required: false
          schema:
            type: string
            enum: [all, pending, validated, rejected, to-review]
        - name: search
          in: query
          description: Recherche dans le titre
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Numéro de page
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Résultats par page
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Liste des réglementations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegulationListResponse'
        '400':
          description: Paramètres invalides
        '401':
          description: Non authentifié

  /regulations/{id}:
    get:
      summary: Détails d'une réglementation
      description: Récupère une réglementation spécifique par son ID
      tags:
        - Réglementations
      parameters:
        - name: id
          in: path
          description: ID de la réglementation
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails de la réglementation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Regulation'
        '404':
          description: Réglementation non trouvée
        '401':
          description: Non authentifié

  /regulations/{id}/status:
    put:
      summary: Mettre à jour le statut
      description: Mettre à jour le statut d'une réglementation (validation/rejet)
      tags:
        - Réglementations
      parameters:
        - name: id
          in: path
          description: ID de la réglementation
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegulationRequest'
      responses:
        '200':
          description: Réglementation mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Regulation'
        '400':
          description: Status invalide
        '404':
          description: Réglementation non trouvée
        '401':
          description: Non authentifié

  /regulations/stats:
    get:
      summary: Statistiques des réglementations
      description: Récupère les statistiques agrégées des réglementations
      tags:
        - Réglementations
      responses:
        '200':
          description: Statistiques
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegulationStats'
        '401':
          description: Non authentifié

components:
  schemas:
    Regulation:
      type: object
      required:
        - id
        - title
        - description
        - status
        - type
        - dateCreated
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique de la réglementation
        title:
          type: string
          description: Titre de la réglementation
        description:
          type: string
          description: Description ou résumé
        status:
          type: string
          enum: [pending, validated, rejected, to-review]
          description: Statut de validation
        type:
          type: string
          description: Type de réglementation
        dateCreated:
          type: string
          format: date-time
          description: Date de création
        reference:
          type: string
          format: uri
          description: URL de référence (optionnel)

    RegulationListResponse:
      type: object
      required:
        - regulations
        - total
        - page
        - limit
      properties:
        regulations:
          type: array
          items:
            $ref: '#/components/schemas/Regulation'
          description: Liste des réglementations
        total:
          type: integer
          description: Nombre total de réglementations
        page:
          type: integer
          description: Page actuelle
        limit:
          type: integer
          description: Résultats par page

    UpdateRegulationRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [validated, rejected, to-review]
          description: Nouveau statut
        comment:
          type: string
          description: Commentaire optionnel

    RegulationStats:
      type: object
      required:
        - total
        - by_status
        - recent_count
        - high_priority
      properties:
        total:
          type: integer
          description: Nombre total de réglementations
        by_status:
          type: object
          properties:
            pending:
              type: integer
            validated:
              type: integer
            rejected:
              type: integer
        recent_count:
          type: integer
          description: Nombre de réglementations récentes
        high_priority:
          type: integer
          description: Nombre de réglementations prioritaires

tags:
  - name: Réglementations
    description: Gestion des réglementations européennes